🎨 Simplify treasury covalent adapters
* Split portfolio and transfers into each their own module.
* Let It Raise whenever an API response is failed.
* Normalize tests a bit.
